Conflicts:
	Projects/Win32Framework/Win32Framework.vcxproj
	Projects/Win32Framework/Win32Framework.vcxproj.filters
	Sources/Internal/Core/Core.cpp
	Sources/Internal/Render/2D/FTFont.cpp
	Sources/Internal/Render/2D/Font.cpp
	Sources/Internal/Render/2D/Sprite.cpp
	Sources/Internal/Render/2D/TextBlock.cpp
	Sources/Internal/Render/2D/TextBlockDistanceRender.cpp
	Sources/Internal/Render/2D/TextBlockSoftwareRender.cpp
	Sources/Internal/UI/UIControlSystem.cpp
	Sources/Internal/UI/UIScreenTransition.cpp
	Sources/Internal/UI/UISlider.cpp

Former-commit-id: 68089004db2ba0095d1d622f685604d865b9cca7

+++ Sources/Internal/Render/2D/RenderSystem2D/RenderSystem2D.cpp
@@ -412,7 +412,7 @@ void RenderSystem2D::PrepareSpriteRenderData(Sprite * sprite, Sprite::DrawState 
         spriteClippedTexCoords.reserve(sprite->clipPolygon->GetPointCount());
         
         Texture * t = sprite->GetTexture(frame);
-        Vector2 virtualTexSize = VirtualCoordinates::ConvertResourceToVirtual(Vector2(t->width, t->height), sprite->GetResourceSizeIndex()) ;
+        Vector2 virtualTexSize = VirtualCoordinates::ConvertResourceToVirtual(Vector2((float32)t->width, (float32)t->height), sprite->GetResourceSizeIndex());
         float32 adjWidth = 1.f / virtualTexSize.x;
         float32 adjHeight = 1.f / virtualTexSize.y;
